requirements.txt
django==5.0.6
psycopg2-binary==2.9.9
djoser==2.2.3
Pillow==10.3.0
djangorestframework==3.15.2